rename to sensu-extensions-influxdb2
