+++ lib/sensu/extensions/influxdb2/version.rb
@@ -1,0 +1,10 @@
+
+module SensuExtensionsInfluxDB2
+  module Version
+    MAJOR = 0
+    MINOR = 0
+    PATCH = 2
+
+    VER_STRING = [MAJOR, MINOR, PATCH].compact.join('.')
+  end
+end
